+++ endpoints/item_conta_estoque.py
@@ -2,6 +2,7 @@ from flask import Blueprint, jsonify, request
 from conexao_db import Conexao
 from logger import logger 
 from login import verifica_token
+from usuario import verifica_acesso
 
 item_conta_estoque_bp = Blueprint('item_conta_estoque', __name__)
 conexao = Conexao()
@@ -9,6 +10,10 @@ conexao = Conexao()
 @item_conta_estoque_bp.route('/itens_conta_estoque', methods=['GET'])
 @verifica_token
 def obter_itens_conta_estoque(payload):
+    id_usuario = payload['id_usuario']
+    if not verifica_acesso(id_usuario, request.method, 'item_conta_estoque'):
+        return jsonify({'message': 'Usuário não possui acesso a consultar Item Conta Estoque'}), 403
+    
     try:
         query = 'SELECT * FROM item_conta_estoque'
         resultado = conexao.execute_query(query)
@@ -26,6 +31,10 @@ def obter_itens_conta_estoque(payload):
 @item_conta_estoque_bp.route('/itens_conta_estoque', methods=['POST'])
 @verifica_token
 def inserir_item_conta_estoque(payload):
+    id_usuario = payload['id_usuario']
+    if not verifica_acesso(id_usuario, request.method, 'item_conta_estoque'):
+        return jsonify({'message': 'Usuário não possui acesso a inserir Item Conta Estoque'}), 403
+    
     try:
         dados_item_conta_estoque = request.get_json()
 
@@ -53,6 +62,10 @@ def inserir_item_conta_estoque(payload):
 @item_conta_estoque_bp.route('/itens_conta_estoque/<int:id_item>/<int:id_conta_estoque>', methods=['PUT'])
 @verifica_token
 def atualizar_item_conta_estoque(payload, id_item, id_conta_estoque):
+    id_usuario = payload['id_usuario']
+    if not verifica_acesso(id_usuario, request.method, 'item_conta_estoque'):
+        return jsonify({'message': 'Usuário não possui acesso a alterar Item Conta Estoque'}), 403
+    
     try:
         dados_item_conta_estoque = request.get_json()
 
@@ -81,6 +94,10 @@ def atualizar_item_conta_estoque(payload, id_item, id_conta_estoque):
 @item_conta_estoque_bp.route('/itens_conta_estoque/<int:id_item>/<int:id_conta_estoque>', methods=['DELETE'])
 @verifica_token
 def deletar_item_conta_estoque(payload, id_item, id_conta_estoque):
+    id_usuario = payload['id_usuario']
+    if not verifica_acesso(id_usuario, request.method, 'item_conta_estoque'):
+        return jsonify({'message': 'Usuário não possui acesso a excluir Item Conta Estoque'}), 403
+    
     try:
         query = "DELETE FROM item_conta_estoque WHERE id_item = %s AND id_conta_estoque = %s"
         params = (id_item, id_conta_estoque)
